test(analytics): execute the live loaders on Postgres, not SQLite
`tests/analytics-live-sqlite.test.ts` built a SQLite database from a
hand-kept fixture and ran the src/analytics-live.ts loaders against it.
That caught what a mock cannot -- the rank CTE, the gap-island window
functions, the HAVING-bound `min_samples` floor only fail at execution
-- but it proved those constructs run on an engine NOTHING SERVES FROM.
src/graphql.ts and src/mcp-server.ts reach these loaders through
`observationsReadDb`, which is Neon, and with D1 deleted there is no
second store a SQLite pass could be evidence about.

Measured, not argued. Restoring the #10086 bug -- `WHERE ok = 0` against
a column that is BOOLEAN on Neon and was INTEGER on D1:

  tests/analytics-live-sqlite.test.ts    17 passed
  tests/no-sqlite-only-sql.test.ts        3 passed
  tests/analytics-live-postgres.test.ts   1 FAILED

Neither the execution suite nor the static gate saw it. That is the gap.

The replacement runs on pglite -- real Postgres compiled to wasm, so no
service and no credential -- through the three seams production uses:

  1. THE SCHEMA IS migrations/neon/*.sql, EXEC'D VERBATIM. Not a fixture
     transliterated from it. A hand-kept copy is a second schema to
     maintain and it fails silently: a column production dropped still
     exists there, so the query reading it passes.
  2. THE ADAPTER IS `pgObservationsReadDb`, the one src/graphql.ts gets.
     Two adapters in this tree present Postgres as a D1-shaped reader
     and they once disagreed about the result envelope; a bespoke double
     would be a third opinion.
  3. THE PLACEHOLDERS GO THROUGH `toPositionalPlaceholders`, the rewrite
     `createPgSql.unsafe` performs, so this cannot pass on a conversion
     production does not do.

Coverage of src/analytics-live.ts is byte-identical to the suite it
replaces (41.97% stmts / 35.71% branch / 56.66% funcs / 45.77% lines),
so nothing was traded for the engine change. One pglite instance per
file with a TRUNCATE between tests keeps it at 1.05s; a fresh instance
per test spent 9 of 10 seconds re-applying identical DDL.

What this is not: pglite is not the Neon deployment. It cannot show a
Hyperdrive behaviour, a pool interaction, or a plan chosen against
production statistics -- those need a throwaway branch, and they are a
different question from "does this SQL execute on Postgres at all",
which is the one that was unanswered.

tests/no-sqlite-only-sql.test.ts still guards the other direction, and
the SQLite fixture stays: two handler suites still build from it.

Closes #10227
